More window number updates from int to NSInteger
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/branches/gnustep_testplant_branch@40212 72102866-910b-0410-8b05-ffd578937521

+++ Headers/x11/XGServerWindow.h
@@ -130,8 +130,8 @@ typedef struct _gswindow_device_t {
 + (gswindow_device_t *) _windowForXWindow: (Window)xWindow;
 + (gswindow_device_t *) _windowForXParent: (Window)xWindow;
 + (gswindow_device_t *) _windowWithTag: (int)windowNumber;
-- (void) _addExposedRectangle: (XRectangle)rectangle : (int)win : (BOOL) ignoreBacking;
-- (void) _processExposedRectangles: (int)win;
+- (void) _addExposedRectangle: (XRectangle)rectangle : (NSInteger)win : (BOOL) ignoreBacking;
+- (void) _processExposedRectangles: (NSInteger)win;
 - (void) _initializeCursorForXWindow: (Window) win;
 - (void) _destroyServerWindows;
